+++ src/PhpSpreadsheet/Writer/Xlsx/Chart.php
@@ -1024,6 +1024,38 @@ class Chart extends WriterPart
     }
 
     /**
+     * Method writing plot series values.
+     *
+     * @param XMLWriter $objWriter XML Writer
+     * @param int       $val       value for idx (default: 3)
+     * @param string    $fillColor hex color (default: FF9900)
+     *
+     * @return XMLWriter XML Writer
+     */
+    private function writePlotSeriesValuesElement($objWriter, $val = 3, $fillColor = 'FF9900')
+    {
+        $objWriter->startElement('c:dPt');
+        $objWriter->startElement('c:idx');
+        $objWriter->writeAttribute('val', $val);
+        $objWriter->endElement();
+
+        $objWriter->startElement('c:bubble3D');
+        $objWriter->writeAttribute('val', 0);
+        $objWriter->endElement();
+
+        $objWriter->startElement('c:spPr');
+        $objWriter->startElement('a:solidFill');
+        $objWriter->startElement('a:srgbClr');
+        $objWriter->writeAttribute('val', $fillColor);
+        $objWriter->endElement();
+        $objWriter->endElement();
+        $objWriter->endElement();
+        $objWriter->endElement();
+
+        return $objWriter;
+    }
+
+    /**
      * Write Plot Group (series of related plots).
      *
      * @param DataSeries $plotGroup
@@ -1078,7 +1110,7 @@ class Chart extends WriterPart
             $plotLabel = $plotGroup->getPlotLabelByIndex($plotSeriesIdx);
             if ($plotLabel) {
                 $fillColor = $plotLabel->getFillColor();
-                if ($fillColor !== null) {
+                if ($fillColor !== null && !is_array($fillColor)) {
                     $objWriter->startElement('c:spPr');
                     $objWriter->startElement('a:solidFill');
                     $objWriter->startElement('a:srgbClr');
@@ -1097,24 +1129,18 @@ class Chart extends WriterPart
             $objWriter->writeAttribute('val', $this->seriesIndex + $plotSeriesRef);
             $objWriter->endElement();
 
+            //    Values
+            $plotSeriesValues = $plotGroup->getPlotValuesByIndex($plotSeriesRef);
+
             if (($groupType == DataSeries::TYPE_PIECHART) || ($groupType == DataSeries::TYPE_PIECHART_3D) || ($groupType == DataSeries::TYPE_DONUTCHART)) {
-                $objWriter->startElement('c:dPt');
-                $objWriter->startElement('c:idx');
-                $objWriter->writeAttribute('val', 3);
-                $objWriter->endElement();
-
-                $objWriter->startElement('c:bubble3D');
-                $objWriter->writeAttribute('val', 0);
-                $objWriter->endElement();
-
-                $objWriter->startElement('c:spPr');
-                $objWriter->startElement('a:solidFill');
-                $objWriter->startElement('a:srgbClr');
-                $objWriter->writeAttribute('val', 'FF9900');
-                $objWriter->endElement();
-                $objWriter->endElement();
-                $objWriter->endElement();
-                $objWriter->endElement();
+                $fillColorValues = $plotSeriesValues->getFillColor();
+                if ($fillColorValues !== null && is_array($fillColorValues)) {
+                    foreach ($plotSeriesValues->getDataValues() as $dataKey => $dataValue) {
+                        $this->writePlotSeriesValuesElement($objWriter, $dataKey, (isset($fillColorValues[$dataKey]) ? $fillColorValues[$dataKey] : 'FF9900'));
+                    }
+                } else {
+                    $this->writePlotSeriesValuesElement($objWriter);
+                }
             }
 
             //    Labels
@@ -1126,9 +1152,6 @@ class Chart extends WriterPart
                 $objWriter->endElement();
                 $objWriter->endElement();
             }
-
-            //    Values
-            $plotSeriesValues = $plotGroup->getPlotValuesByIndex($plotSeriesRef);
 
             //    Formatting for the points
             if (($groupType == DataSeries::TYPE_LINECHART) || ($groupType == DataSeries::TYPE_STOCKCHART)) {